Revert "Initialize conf.cfg inside sdk.InitStorageSDK if it is possible"
