+++ core/transaction/vars.go
@@ -2,10 +2,21 @@ package transaction
 
 import (
 	"errors"
+
+	"github.com/0chain/gosdk/core/conf"
 )
 
 var (
-	// ErrInvalidRequest invalid request
+	// Config
+	cfg *conf.Config
+)
+
+// SetConfig set config variables for transaction
+func SetConfig(c *conf.Config) {
+	cfg = c
+}
+
+var (
 	ErrInvalidRequest = errors.New("[txn] invalid request")
 
 	// ErrNoAvailableSharder no any available sharder
@@ -16,4 +27,7 @@ var (
 
 	// ErrTooLessConfirmation too less sharder to confirm transaction
 	ErrTooLessConfirmation = errors.New("[txn] too less sharders to confirm it")
+
+	// ErrConfigIsNotInitialized config is not initialized
+	ErrConfigIsNotInitialized = errors.New("[txn] config is not initialized. please initialize it by transaction.SetConfig")
 )
