+++ zboxcore/sdk/networkworker.go
@@ -13,7 +13,6 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/0chain/errors"
-	"github.com/0chain/gosdk/core/conf"
 	"github.com/0chain/gosdk/zboxcore/blockchain"
 	"github.com/0chain/gosdk/zboxcore/zboxutil"
 )
@@ -56,10 +55,6 @@ func UpdateNetworkDetails() error {
 		sdkInitialized = false
 		blockchain.SetMiners(networkDetails.Miners)
 		blockchain.SetSharders(networkDetails.Sharders)
-		conf.InitChainNetwork(&conf.Network{
-			Sharders: networkDetails.Sharders,
-			Miners:   networkDetails.Miners,
-		})
 		sdkInitialized = true
 	}
 	return nil
