+++ zboxcore/sdk/sdk.go
@@ -76,28 +76,17 @@ func GetLogger() *logger.Logger {
 	return &Logger
 }
 
-// InitStorageSDK init storage sdk with walletJSON
-//   {
-//		"client_id":"322d1dadec182effbcbdeef77d84f",
-//		"client_key":"3b6d02a22ec82d4d9aa1402917ca2",
-//		"keys":[{
-//			"public_key":"3b6d02a22ec82d4d9aa1402917ca268",
-//			"private_key":"25f2e1355d3864de01aba0bfec3702"
-//			}],
-//		"mnemonics":"double wink spin mushroom thing notable trumpet chapter",
-//		"version":"1.0",
-//		"date_created":"2021-08-18T08:34:39+08:00"
-//	 }
-func InitStorageSDK(walletJSON string, blockWorker, chainID, signatureScheme string, preferredBlobbers []string) error {
-
-	err := client.PopulateClient(walletJSON, signatureScheme)
+// InitStorageSDK init storage sdk with wallet json and config
+func InitStorageSDK(walletJSON string, cfg conf.Config) error {
+	err := client.PopulateClient(walletJSON, cfg.SignatureScheme)
 	if err != nil {
 		return err
 	}
+	blockchain.SetChainID(cfg.ChainID)
+	blockchain.SetPreferredBlobbers(cfg.PreferredBlobbers)
+	blockchain.SetBlockWorker(cfg.BlockWorker)
 
-	blockchain.SetChainID(chainID)
-	blockchain.SetPreferredBlobbers(preferredBlobbers)
-	blockchain.SetBlockWorker(blockWorker)
+	transaction.SetConfig(&cfg)
 
 	err = UpdateNetworkDetails()
 	if err != nil {
@@ -118,22 +107,10 @@ func GetNetwork() *Network {
 
 func SetMaxTxnQuery(num int) {
 	blockchain.SetMaxTxnQuery(num)
-
-	cfg, _ := conf.GetClientConfig()
-	if cfg != nil {
-		cfg.MaxTxnQuery = num
-	}
-
 }
 
 func SetQuerySleepTime(time int) {
 	blockchain.SetQuerySleepTime(time)
-
-	cfg, _ := conf.GetClientConfig()
-	if cfg != nil {
-		cfg.QuerySleepTime = time
-	}
-
 }
 
 func SetMinSubmit(num int) {
